Adding extra files with useful code for potential future use
